move stories to separate folder
