+++ .storybook/config.js
@@ -5,8 +5,7 @@ import { ThemeProvider } from 'styled-components';
 import '../src/global-styles';
 import def from '../src/theme/light';
 
-// automatically import all files ending in *.stories.js
-const req = require.context('../src/components', true, /.stories.tsx$/);
+const req = require.context('../src/stories', true, /.stories.tsx$/);
 function loadStories() {
   req.keys().forEach(filename => req(filename));
 }
