+++ src/stories/ModalOkCancel.stories.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import ModalOkCancel from './ModalOkCancel';
+import ModalOkCancel from '../components/ModalOkCancel';
 
 const stories = storiesOf('ModalOkCancel', module);
