+++ src/stories/TextButton.stories.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import TextButton from './TextButton';
+import TextButton from '../components/TextButton';
 
 const stories = storiesOf('Button', module);
