+++ src/stories/TextInput.stories.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import TextInput from './TextInput';
+import TextInput from '../components/TextInput';
 
 const stories = storiesOf('TextInput', module);
